types cleanup

+++ Emulator/FileSystems/FSBitmapBlock.cpp
@@ -42,7 +42,7 @@ FSBitmapBlock::dump() const
     u32 count = 0;
     for (u32 i = 1; i < bsize() / 4; i++) {
         if (u32 value = get32(i)) {
-            for (int j = 0; j < 32; j++) {
+            for (usize j = 0; j < 32; j++) {
                 if (GET_BIT(value, j)) count++;
             }
         }
