+++ Emulator/Files/DiskFiles/IMGFile.cpp
@@ -137,7 +137,7 @@ IMGFile::encodeTrack(Disk *disk, Track t)
 
     // Encode track header
     p += 82;                                        // GAP
-    for (int i = 0; i < 24; i++) { p[i] = 0xAA; }   // SYNC
+    for (usize i = 0; i < 24; i++) { p[i] = 0xAA; } // SYNC
     p += 24;
     p[0] = 0x52; p[1] = 0x24;                       // IAM
     p[2] = 0x52; p[3] = 0x24;
@@ -166,7 +166,7 @@ IMGFile::encodeSector(Disk *disk, Track t, Sector s)
     debug(MFM_DEBUG, "  Encoding DOS sector %d\n", s);
     
     // Write SYNC
-    for (int i = 0; i < 12; i++) { buf[i] = 0x00; }
+    for (usize i = 0; i < 12; i++) { buf[i] = 0x00; }
     
     // Write IDAM
     buf[12] = 0xA1;
@@ -186,10 +186,10 @@ IMGFile::encodeSector(Disk *disk, Track t, Sector s)
     buf[21] = LO_BYTE(crc);
 
     // Write GAP
-    for (int i = 22; i < 44; i++) { buf[i] = 0x4E; }
+    for (usize i = 22; i < 44; i++) { buf[i] = 0x4E; }
 
     // Write SYNC
-    for (int i = 44; i < 56; i++) { buf[i] = 0x00; }
+    for (usize i = 44; i < 56; i++) { buf[i] = 0x00; }
 
     // Write DATA AM
     buf[56] = 0xA1;
@@ -270,7 +270,7 @@ IMGFile::decodeTrack(Disk *disk, Track t)
     
     // Determine the start of all sectors contained in this track
     int sectorStart[numSectors];
-    for (int i = 0; i < numSectors; i++) {
+    for (isize i = 0; i < numSectors; i++) {
         sectorStart[i] = 0;
     }
     int cnt = 0;
@@ -312,7 +312,7 @@ IMGFile::decodeTrack(Disk *disk, Track t)
     }
         
     // Do some consistency checking
-    for (int i = 0; i < numSectors; i++) assert(sectorStart[i] != 0);
+    for (isize i = 0; i < numSectors; i++) assert(sectorStart[i] != 0);
     
     // Encode all sectors
     bool result = true;
